requirements.txt
boto3>=1.26.0
faker>=19.0.0
pandas>=1.5.0
pyarrow>=10.0.0
python-dotenv>=1.0.0
apache-airflow>=2.7.0
pyiceberg>=0.5.0
pyspark>=3.3.0
mlflow>=2.8.0
apache-airflow-providers-fab